Added GuiRearrangeHelper (#1678)
* added GuiRearrangeHelper; shared GuiRearrangeHelper between AppStateRearrange and AppStateSocialNavStudy
* add GuiRearrangeHelper.update method
* add GuiRearrangeHelper docstring
